ADD Refactor the device registry to allow multiple registries
