+++ lib/errors.js
@@ -27,16 +27,17 @@ module.exports = {
     RegistrationError: function(id, type) {
         this.name = 'REGISTRATION_ERROR';
         this.message = 'Error registering context provider for device: ' + id + ' of type: ' + type;
-        this.code = '500';
     },
     UnregistrationError: function(id, type) {
         this.name = 'UNREGISTRATION_ERROR';
         this.message = 'Error unregistering context provider for device: ' + id + ' of type: ' + type;
-        this.code = '500';
     },
     EntityUpdateError: function(id, type) {
         this.name = 'ENTITY_UPDATE_ERROR';
         this.message = 'Error updating entity data for device: ' + id + ' of type: ' + type;
-        this.code = '500';
+    },
+    RegistryNotAvailable: function() {
+        this.name = 'REGISTRY_NOT_AVAILABLE';
+        this.message = 'No device registry is available.';
     }
 };
