+++ lib/fiware-iotagent-lib.js
@@ -26,19 +26,22 @@
 var request = require('request'),
     async = require('async'),
     errors = require('./errors'),
-    registry = require('./services/deviceRegistry'),
     utils = require('./asyncUtils'),
     contextServer = require('./services/contextServer'),
     logger = require('fiware-node-logger'),
+    registry,
     config;
 
 /**
- * Activates the IoT Agent to start listening for NGSI Calls (to act as a Context Provider).
+ * Activates the IoT Agent to start listening for NGSI Calls (to act as a Context Provider). It also creates the
+ * device registry for the IoT Agent (based on the deviceRegistry.type configuration option).
  *
  * @param {Object} newConfig            Configuration of the Context Server
  */
 function activate(newConfig, callback) {
     config = newConfig;
+
+    registry = require('./services/deviceRegistryMemory');
 
     async.series([
         async.apply(contextServer.start, newConfig)
@@ -240,11 +243,20 @@ function setDataQueryHandler(dataHandler) {
     contextServer.setQueryHandler(dataHandler);
 }
 
+function listDevices(callback) {
+    if (registry) {
+        registry.list(callback);
+    } else {
+        logger.error('Tried to list devices before a registry was available');
+        callback(new errors.RegistryNotAvailable());
+    }
+}
+
 exports.activate = activate;
 exports.deactivate = deactivate;
 exports.register = registerDevice;
 exports.unregister = unregisterDevice;
 exports.update = updateValue;
-exports.listDevices = registry.list;
+exports.listDevices = listDevices;
 exports.setDataUpdateHandler = setDataUpdateHandler;
 exports.setDataQueryHandler = setDataQueryHandler;
