+++ lib/services/deviceRegistryMemory.js
